parse_text {spelling}	R Documentation

Text Parsers

Description

Parse text from various formats and return a data frame with text lines and position in the source document.

Usage

parse_text_md(path, extensions = TRUE)

Arguments

path
markdown file

extensions
render markdown extensions? Passed to commonmark

[Package spelling version 1.2 Index]