spell_check_files {spelling}	R Documentation

Spell Check

Description

Perform a spell check on document files or plain text.

Usage

spell_check_files(path, ignore = character(), lang = "en_US")

spell_check_text(text, ignore = character(), lang = "en_US")

Arguments

path
path to file or to spell check

ignore
character vector with words which will be added to the hunspell::dictionary

lang
set Language field in DESCRIPTION e.g. "en-US" or "en-GB".

text
character vector with plain text

See Also

Other spelling: spell_check_package, wordlist

Examples

# Example files
files <- list.files(system.file("examples", package = "knitr"),
  pattern = "\\.(Rnw|Rmd|html)$", full.names = TRUE)
spell_check_files(files)

[Package spelling version 1.2 Index]